+++ msg/scratch_msgs.js
@@ -306,304 +306,4 @@ Blockly.ScratchMsgs.locales["en"] =
     "BROADCAST_MODAL_TITLE": "New Message",
     "DEFAULT_BROADCAST_MESSAGE_NAME": "message1"
 };
-
-Blockly.ScratchMsgs.locales["zh-cn"] =
-{
-    "CONTROL_FOREVER": "重复执行",
-    "CONTROL_REPEAT": "重复执行 %1 次",
-    "CONTROL_IF": "如果 %1 那么",
-    "CONTROL_ELSE": "否则",
-    "CONTROL_STOP": "停止",
-    "CONTROL_STOP_ALL": "全部脚本",
-    "CONTROL_STOP_THIS": "这个脚本",
-    "CONTROL_STOP_OTHER": "该角色的其他脚本",
-    "CONTROL_WAIT": "等待 %1 秒",
-    "CONTROL_WAITUNTIL": "等待 %1",
-    "CONTROL_REPEATUNTIL": "重复执行直到 %1",
-    "CONTROL_WHILE": "当 %1 重复执行",
-    "CONTROL_FOREACH": "对于 %2 中的每个 %1",
-    "CONTROL_STARTASCLONE": "当作为克隆体启动时",
-    "CONTROL_CREATECLONEOF": "克隆 %1",
-    "CONTROL_CREATECLONEOF_MYSELF": "自己",
-    "CONTROL_DELETETHISCLONE": "删除此克隆体",
-    "CONTROL_COUNTER": "计数器",
-    "CONTROL_INCRCOUNTER": "计数器加一",
-    "CONTROL_CLEARCOUNTER": "计数器归零",
-    "CONTROL_ALLATONCE": "所有脚本",
-    "DATA_SETVARIABLETO": "将 %1 设为 %2",
-    "DATA_CHANGEVARIABLEBY": "将 %1 增加 %2",
-    "DATA_SHOWVARIABLE": "显示变量 %1",
-    "DATA_HIDEVARIABLE": "隐藏变量 %1",
-    "DATA_ADDTOLIST": "将 %1 加入 %2",
-    "DATA_DELETEOFLIST": "删除 %2 的第 %1 项",
-    "DATA_DELETEALLOFLIST": "删除 %1 的全部项目",
-    "DATA_INSERTATLIST": "在 %3 的第 %2 项前插入 %1",
-    "DATA_REPLACEITEMOFLIST": "将 %2 的第 %1 项替换为 %3",
-    "DATA_ITEMOFLIST": "%2 的第 %1 项",
-    "DATA_ITEMNUMOFLIST": "%2 中第一个 %1 的编号",
-    "DATA_LENGTHOFLIST": "%1 的项目数",
-    "DATA_LISTCONTAINSITEM": "%1 包含 %2 ?",
-    "DATA_SHOWLIST": "显示列表 %1",
-    "DATA_HIDELIST": "隐藏列表 %1",
-    "DATA_INDEX_ALL": "全部",
-    "DATA_INDEX_LAST": "末尾",
-    "DATA_INDEX_RANDOM": "随机",
-    "EVENT_WHENARDUINOBEGIN": "当Arduino启动",
-    "EVENT_WHENMICROBITBEGIN": "当Microbit启动",
-    "EVENT_WHENMICROBITBUTTONPRESSED": "当按键 %1 被按下",
-    "EVENT_WHENMICROBITPINBEINGTOUCHED": "当引脚 %1 被触摸",
-    "EVENT_WHENMICROBITGESTURE": "当姿态 %1",
-    "EVENT_WHENFLAGCLICKED": "当 %1 被点击",
-    "EVENT_WHENTHISSPRITECLICKED": "当角色被点击",
-    "EVENT_WHENSTAGECLICKED": "当舞台被点击",
-    "EVENT_WHENTOUCHINGOBJECT": "当该角色碰到 %1",
-    "EVENT_WHENBROADCASTRECEIVED": "当接收到 %1",
-    "EVENT_WHENBACKDROPSWITCHESTO": "当背景换成 %1",
-    "EVENT_WHENGREATERTHAN": "当 %1 > %2",
-    "EVENT_WHENGREATERTHAN_TIMER": "计时器",
-    "EVENT_WHENGREATERTHAN_LOUDNESS": "响度",
-    "EVENT_BROADCAST": "广播 %1",
-    "EVENT_BROADCASTANDWAIT": "广播 %1 并等待",
-    "EVENT_WHENKEYPRESSED": "当按下 %1 键",
-    "EVENT_WHENMICROBITGESTURE_SHAKEN": "摇晃",
-    "EVENT_WHENMICROBITGESTURE_TILTEDUPWARD": "向上倾斜",
-    "EVENT_WHENMICROBITGESTURE_TILTEDDOWNWARD": "向下倾斜",
-    "EVENT_WHENMICROBITGESTURE_TILTEDDLEFTWARD": "向左倾斜",
-    "EVENT_WHENMICROBITGESTURE_TILTEDDRIGHTWARD": "向右倾斜",
-    "EVENT_WHENMICROBITGESTURE_FACEUP": "正面朝上",
-    "EVENT_WHENMICROBITGESTURE_FACEDOWN": "正面朝下",
-    "EVENT_WHENMICROBITGESTURE_FREEFALL": "自由落体",
-    "EVENT_WHENKEYPRESSED_SPACE": "空格",
-    "EVENT_WHENKEYPRESSED_LEFT": "←",
-    "EVENT_WHENKEYPRESSED_RIGHT": "→",
-    "EVENT_WHENKEYPRESSED_DOWN": "↓",
-    "EVENT_WHENKEYPRESSED_UP": "↑",
-    "EVENT_WHENKEYPRESSED_ANY": "任意",
-    "LOOKS_SAYFORSECS": "说 %1 %2 秒",
-    "LOOKS_SAY": "说 %1",
-    "LOOKS_HELLO": "你好！",
-    "LOOKS_THINKFORSECS": "思考 %1 %2 秒",
-    "LOOKS_THINK": "思考 %1",
-    "LOOKS_HMM": "嗯……",
-    "LOOKS_SHOW": "显示",
-    "LOOKS_HIDE": "隐藏",
-    "LOOKS_HIDEALLSPRITES": "隐藏所有角色",
-    "LOOKS_EFFECT_COLOR": "颜色",
-    "LOOKS_EFFECT_FISHEYE": "鱼眼",
-    "LOOKS_EFFECT_WHIRL": "漩涡",
-    "LOOKS_EFFECT_PIXELATE": "像素化",
-    "LOOKS_EFFECT_MOSAIC": "马赛克",
-    "LOOKS_EFFECT_BRIGHTNESS": "亮度",
-    "LOOKS_EFFECT_GHOST": "虚像",
-    "LOOKS_CHANGEEFFECTBY": "将 %1 特效增加 %2",
-    "LOOKS_SETEFFECTTO": "将 %1 特效设定为 %2",
-    "LOOKS_CLEARGRAPHICEFFECTS": "清除图形特效",
-    "LOOKS_CHANGESIZEBY": "将大小增加 %1",
-    "LOOKS_SETSIZETO": "将大小设为 %1",
-    "LOOKS_SIZE": "大小",
-    "LOOKS_CHANGESTRETCHBY": "伸缩%1",
-    "LOOKS_SETSTRETCHTO": "设置伸缩为%1 %",
-    "LOOKS_SWITCHCOSTUMETO": "换成 %1 造型",
-    "LOOKS_NEXTCOSTUME": "下一个造型",
-    "LOOKS_SWITCHBACKDROPTO": "换成 %1 背景",
-    "LOOKS_GOTOFRONTBACK": "移到最 %1 ",
-    "LOOKS_GOTOFRONTBACK_FRONT": "前面",
-    "LOOKS_GOTOFRONTBACK_BACK": "后面",
-    "LOOKS_GOFORWARDBACKWARDLAYERS": "%1 %2 层",
-    "LOOKS_GOFORWARDBACKWARDLAYERS_FORWARD": "前移",
-    "LOOKS_GOFORWARDBACKWARDLAYERS_BACKWARD": "后移",
-    "LOOKS_BACKDROPNUMBERNAME": "背景 %1",
-    "LOOKS_COSTUMENUMBERNAME": "造型 %1",
-    "LOOKS_NUMBERNAME_NUMBER": "编号",
-    "LOOKS_NUMBERNAME_NAME": "名称",
-    "LOOKS_SWITCHBACKDROPTOANDWAIT": "换成 %1 背景并等待",
-    "LOOKS_NEXTBACKDROP_BLOCK": "下一个背景",
-    "LOOKS_NEXTBACKDROP": "下一个背景",
-    "LOOKS_PREVIOUSBACKDROP": "上一个背景",
-    "LOOKS_RANDOMBACKDROP": "随机背景",
-    "MOTION_MOVESTEPS": "移动 %1 步",
-    "MOTION_TURNLEFT": "左转 %1 %2 度",
-    "MOTION_TURNRIGHT": "右转 %1 %2 度",
-    "MOTION_POINTINDIRECTION": "面向 %1 方向",
-    "MOTION_POINTTOWARDS": "面向 %1",
-    "MOTION_POINTTOWARDS_POINTER": "鼠标指针",
-    "MOTION_POINTTOWARDS_RANDOM": "随机方向",
-    "MOTION_GOTO": "移到 %1",
-    "MOTION_GOTO_POINTER": "鼠标指针",
-    "MOTION_GOTO_RANDOM": "随机位置",
-    "MOTION_GOTOXY": "移到 x: %1 y: %2",
-    "MOTION_GLIDESECSTOXY": "在 %1 秒内滑行到 x: %2 y: %3",
-    "MOTION_GLIDETO": "在 %1 秒内滑行到 %2",
-    "MOTION_GLIDETO_POINTER": "鼠标指针",
-    "MOTION_GLIDETO_RANDOM": "随机位置",
-    "MOTION_CHANGEXBY": "将x坐标增加 %1",
-    "MOTION_SETX": "将x坐标设为 %1",
-    "MOTION_CHANGEYBY": "将y坐标增加 %1",
-    "MOTION_SETY": "将y坐标设为 %1",
-    "MOTION_IFONEDGEBOUNCE": "碰到边缘就反弹",
-    "MOTION_SETROTATIONSTYLE": "将旋转方式设为 %1",
-    "MOTION_SETROTATIONSTYLE_LEFTRIGHT": "左右翻转",
-    "MOTION_SETROTATIONSTYLE_DONTROTATE": "不可旋转",
-    "MOTION_SETROTATIONSTYLE_ALLAROUND": "任意旋转",
-    "MOTION_XPOSITION": "x 坐标",
-    "MOTION_YPOSITION": "y 坐标",
-    "MOTION_DIRECTION": "方向",
-    "MOTION_SCROLLRIGHT": "向右滚动 %1",
-    "MOTION_SCROLLUP": "向上滚动 %1",
-    "MOTION_ALIGNSCENE": "和场景 %1 对齐",
-    "MOTION_ALIGNSCENE_BOTTOMLEFT": "左下角",
-    "MOTION_ALIGNSCENE_BOTTOMRIGHT": "右下角",
-    "MOTION_ALIGNSCENE_MIDDLE": "中间",
-    "MOTION_ALIGNSCENE_TOPLEFT": "左上角",
-    "MOTION_ALIGNSCENE_TOPRIGHT": "右上角",
-    "MOTION_XSCROLL": "x滚动位置",
-    "MOTION_YSCROLL": "y滚动位置",
-    "MOTION_STAGE_SELECTED": "选中了舞台：不可使用运动类积木",
-    "OPERATORS_ADD": "%1 + %2",
-    "OPERATORS_SUBTRACT": "%1 - %2",
-    "OPERATORS_MULTIPLY": "%1 * %2",
-    "OPERATORS_DIVIDE": "%1 / %2",
-    "OPERATORS_RANDOM": "在 %1 和 %2 之间取随机数",
-    "OPERATORS_GT": "%1 > %2",
-    "OPERATORS_LT": "%1 < %2",
-    "OPERATORS_EQUALS": "%1 = %2",
-    "OPERATORS_AND": "%1 与 %2",
-    "OPERATORS_OR": "%1 或 %2",
-    "OPERATORS_NOT": "%1 不成立",
-    "OPERATORS_JOIN": "连接 %1 和 %2",
-    "OPERATORS_JOIN_APPLE": "苹果",
-    "OPERATORS_JOIN_BANANA": "香蕉",
-    "OPERATORS_LETTEROF": "%2 的第 %1 个字符",
-    "OPERATORS_LETTEROF_APPLE": "果",
-    "OPERATORS_LENGTH": "%1 的字符数",
-    "OPERATORS_CONTAINS": "%1 包含 %2 ?",
-    "OPERATORS_MOD": "%1 除以 %2 的余数",
-    "OPERATORS_ROUND": "四舍五入 %1",
-    "OPERATORS_MATHOP": "%1 %2",
-    "OPERATORS_MATHOP_ABS": "绝对值",
-    "OPERATORS_MATHOP_FLOOR": "向下取整",
-    "OPERATORS_MATHOP_CEILING": "向上取整",
-    "OPERATORS_MATHOP_SQRT": "平方根",
-    "OPERATORS_MATHOP_SIN": "sin",
-    "OPERATORS_MATHOP_COS": "cos",
-    "OPERATORS_MATHOP_TAN": "tan",
-    "OPERATORS_MATHOP_ASIN": "asin",
-    "OPERATORS_MATHOP_ACOS": "acos",
-    "OPERATORS_MATHOP_ATAN": "atan",
-    "OPERATORS_MATHOP_LN": "ln",
-    "OPERATORS_MATHOP_LOG": "log",
-    "OPERATORS_MATHOP_EEXP": "e ^",
-    "OPERATORS_MATHOP_10EXP": "10 ^",
-    "PROCEDURES_DEFINITION": "定义 %1",
-    "SENSING_TOUCHINGOBJECT": "碰到 %1 ?",
-    "SENSING_TOUCHINGOBJECT_POINTER": "鼠标指针",
-    "SENSING_TOUCHINGOBJECT_EDGE": "舞台边缘",
-    "SENSING_TOUCHINGCOLOR": "碰到颜色 %1 ?",
-    "SENSING_COLORISTOUCHINGCOLOR": "颜色 %1 碰到 %2 ?",
-    "SENSING_DISTANCETO": "到 %1 的距离",
-    "SENSING_DISTANCETO_POINTER": "鼠标指针",
-    "SENSING_ASKANDWAIT": "询问 %1 并等待",
-    "SENSING_ASK_TEXT": "你叫什么名字？",
-    "SENSING_ANSWER": "回答",
-    "SENSING_KEYPRESSED": "按下 %1 键?",
-    "SENSING_MOUSEDOWN": "按下鼠标?",
-    "SENSING_MOUSEX": "鼠标的x坐标",
-    "SENSING_MOUSEY": "鼠标的y坐标",
-    "SENSING_SETDRAGMODE": "将拖动模式设为 %1",
-    "SENSING_SETDRAGMODE_DRAGGABLE": "可拖动",
-    "SENSING_SETDRAGMODE_NOTDRAGGABLE": "不可拖动",
-    "SENSING_LOUDNESS": "响度",
-    "SENSING_LOUD": "响声？",
-    "SENSING_TIMER": "计时器",
-    "SENSING_RESETTIMER": "计时器归零",
-    "SENSING_OF": "%2 的 %1",
-    "SENSING_OF_XPOSITION": "x 坐标",
-    "SENSING_OF_YPOSITION": "y 坐标",
-    "SENSING_OF_DIRECTION": "方向",
-    "SENSING_OF_COSTUMENUMBER": "造型编号",
-    "SENSING_OF_COSTUMENAME": "造型名称",
-    "SENSING_OF_SIZE": "大小",
-    "SENSING_OF_VOLUME": "音量",
-    "SENSING_OF_BACKDROPNUMBER": "背景编号",
-    "SENSING_OF_BACKDROPNAME": "背景名称",
-    "SENSING_OF_STAGE": "舞台",
-    "SENSING_CURRENT": "当前时间的 %1",
-    "SENSING_CURRENT_YEAR": "年",
-    "SENSING_CURRENT_MONTH": "月",
-    "SENSING_CURRENT_DATE": "日",
-    "SENSING_CURRENT_DAYOFWEEK": "星期",
-    "SENSING_CURRENT_HOUR": "时",
-    "SENSING_CURRENT_MINUTE": "分",
-    "SENSING_CURRENT_SECOND": "秒",
-    "SENSING_DAYSSINCE2000": "2000年至今的天数",
-    "SENSING_USERNAME": "用户名",
-    "SENSING_USERID": "用户id",
-    "SOUND_PLAY": "播放声音 %1",
-    "SOUND_PLAYUNTILDONE": "播放声音 %1 等待播完",
-    "SOUND_STOPALLSOUNDS": "停止所有声音",
-    "SOUND_SETEFFECTO": "将 %1 音效设为 %2",
-    "SOUND_CHANGEEFFECTBY": "将 %1 音效增加 %2",
-    "SOUND_CLEAREFFECTS": "清除音效",
-    "SOUND_EFFECTS_PITCH": "音调",
-    "SOUND_EFFECTS_PAN": "左右平衡",
-    "SOUND_CHANGEVOLUMEBY": "将音量增加 %1",
-    "SOUND_SETVOLUMETO": "将音量设为 %1%",
-    "SOUND_VOLUME": "音量",
-    "SOUND_RECORD": "录制…",
-    "CATEGORY_MOTION": "运动",
-    "CATEGORY_LOOKS": "外观",
-    "CATEGORY_SOUND": "声音",
-    "CATEGORY_EVENTS": "事件",
-    "CATEGORY_CONTROL": "控制",
-    "CATEGORY_SENSING": "侦测",
-    "CATEGORY_OPERATORS": "运算",
-    "CATEGORY_VARIABLES": "变量",
-    "CATEGORY_MYBLOCKS": "自制积木",
-    "DUPLICATE": "复制",
-    "DELETE": "删除",
-    "ADD_COMMENT": "添加注释",
-    "REMOVE_COMMENT": "删除注释",
-    "DELETE_BLOCK": "删除",
-    "DELETE_X_BLOCKS": "删除 %1 积木",
-    "DELETE_ALL_BLOCKS": "删除全部 %1 积木？",
-    "CLEAN_UP": "整理积木",
-    "HELP": "帮助",
-    "UNDO": "撤销",
-    "REDO": "重做",
-    "EDIT_PROCEDURE": "编辑",
-    "SHOW_PROCEDURE_DEFINITION": "查看定义",
-    "WORKSPACE_COMMENT_DEFAULT_TEXT": "说些什么……",
-    "COLOUR_HUE_LABEL": "颜色",
-    "COLOUR_SATURATION_LABEL": "饱和度",
-    "COLOUR_BRIGHTNESS_LABEL": "亮度",
-    "CHANGE_VALUE_TITLE": "更改变量：",
-    "RENAME_VARIABLE": "修改变量名",
-    "RENAME_VARIABLE_TITLE": "将所有的「%1」变量名改为：",
-    "RENAME_VARIABLE_MODAL_TITLE": "修改变量名",
-    "NEW_VARIABLE": "建立一个变量",
-    "NEW_VARIABLE_TITLE": "新变量名：",
-    "VARIABLE_MODAL_TITLE": "新建变量",
-    "VARIABLE_ALREADY_EXISTS": "已经存在名为「%1」的变量。",
-    "VARIABLE_ALREADY_EXISTS_FOR_ANOTHER_TYPE": "已经存在一个名为「%1」的变量，其类型为「%2」。",
-    "DELETE_VARIABLE_CONFIRMATION": "删除%1处「%2」变量吗？",
-    "CANNOT_DELETE_VARIABLE_PROCEDURE": "无法删除变量「%1」，因为函数「%2」的定义中用到了它",
-    "DELETE_VARIABLE": "删除变量「%1」",
-    "NEW_PROCEDURE": "制作新的积木",
-    "PROCEDURE_ALREADY_EXISTS": "已经存在名为「%1」的程序。",
-    "PROCEDURE_DEFAULT_NAME": "积木名称",
-    "PROCEDURE_USED": "To delete a block definition, first remove all uses of the block",
-    "NEW_LIST": "建立一个列表",
-    "NEW_LIST_TITLE": "新的列表名：",
-    "LIST_MODAL_TITLE": "新建列表",
-    "LIST_ALREADY_EXISTS": "名为 「%1」 的列表已存在。",
-    "RENAME_LIST_TITLE": "将所有的「%1」列表改名为：",
-    "RENAME_LIST_MODAL_TITLE": "修改列表名",
-    "DEFAULT_LIST_ITEM": "东西",
-    "DELETE_LIST": "删除「%1」列表",
-    "RENAME_LIST": "修改列表名",
-    "NEW_BROADCAST_MESSAGE": "新消息",
-    "NEW_BROADCAST_MESSAGE_TITLE": "新消息的名称：",
-    "BROADCAST_MODAL_TITLE": "新消息",
-    "DEFAULT_BROADCAST_MESSAGE_NAME": "消息1"
-};
 // End of combined translations
